+++ src/main/java/nl/dtls/fairdatapoint/api/controller/index/AdminController.java
@@ -22,7 +22,7 @@
  */
 package nl.dtls.fairdatapoint.api.controller.index;
 
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.extern.log4j.Log4j2;
 import nl.dtls.fairdatapoint.api.dto.index.ping.PingDTO;
 import nl.dtls.fairdatapoint.entity.index.event.Event;
@@ -52,7 +52,7 @@ public class AdminController {
     @Autowired
     private WebhookService webhookService;
 
-    @ApiOperation(value = "trigger", hidden = true)
+    @Operation(hidden = true)
     @PostMapping("/trigger")
     @PreAuthorize("hasRole('ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
@@ -63,7 +63,7 @@ public class AdminController {
         eventService.triggerMetadataRetrieval(event);
     }
 
-    @ApiOperation(value = "trigger-all", hidden = true)
+    @Operation(hidden = true)
     @PostMapping("/trigger-all")
     @PreAuthorize("hasRole('ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
@@ -74,7 +74,7 @@ public class AdminController {
         eventService.triggerMetadataRetrieval(event);
     }
 
-    @ApiOperation(value = "ping webhook", hidden = true)
+    @Operation(hidden = true)
     @PostMapping("/ping-webhook")
     @PreAuthorize("hasRole('ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)