+++ CHANGELOG.md
@@ -16,6 +16,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Migrated API Docs to SpringDoc OpenAPI
 - Allow to configure multiple ping endpoints
+- Upgrade to Spring Boot 2.4.5
+- Migrated from Mongobee to Mongock
 
 ## [1.8.0]